feat: authorization

+++ src/controllers/appointmentListPatientController.js
@@ -3,8 +3,20 @@ import asyncHandler from "express-async-handler";
 import sequelize from "sequelize";
 const getAllAppointmentList = asyncHandler(async (req, res) => {
   try {
-     if (req.isAuthenticated()) {
+    if (req.isAuthenticated()) {
       try {
+        const userGroup = await db.userGroup.findOne({
+          where: { groupName: req.user.user.role },
+        });
+        const authorization = await db.authorizations.findOne({
+          where: { userGroupId: userGroup.id, featId: 3 },
+        });
+        if (!authorization.isAccess) {
+          return res.status(401).json({
+            status: res.statusCode,
+            message: "Unauthorized",
+          });
+        }
         const appointmentList = await db.appointmentListPatient.findAll({
           attributes: {
             include: [
@@ -46,13 +58,13 @@ const getAllAppointmentList = asyncHandler(async (req, res) => {
           data: "",
         });
       }
-     } else {
-       res.status(401).json({
-         status: res.statusCode,
-         message: "Unauthorized",
-         data: "",
-       });
-     }
+    } else {
+      res.status(401).json({
+        status: res.statusCode,
+        message: "Unauthorized",
+        data: "",
+      });
+    }
   } catch (err) {
     res.status(500).json({
       status: res.statusCode,
@@ -63,7 +75,15 @@ const getAllAppointmentList = asyncHandler(async (req, res) => {
 });
 const createAppointmentListPatient = asyncHandler(async (req, res) => {
   try {
-     if (req.isAuthenticated()) {
+    if (req.isAuthenticated()) {
+      const userGroup = await db.userGroup.findOne({where:{groupName: req.user.user.role}});
+      const authorization = await db.authorizations.findOne({where:{userGroupId: userGroup.id, featId: 4}});
+      if (!authorization.isAccess) { 
+        return res.status(401).json({
+          status: res.statusCode,
+          message: "Unauthorized",
+        });
+      } 
       const { patientId, appointmentListId } = req.body;
       if (!patientId || !appointmentListId) {
         res.status(400).json({
@@ -81,7 +101,7 @@ const createAppointmentListPatient = asyncHandler(async (req, res) => {
         message: "New appointmentList created",
         data: appointmentListPatient,
       });
-     }
+    }
   } catch (err) {
     res.status(500).json({
       status: res.statusCode,
@@ -92,7 +112,15 @@ const createAppointmentListPatient = asyncHandler(async (req, res) => {
 });
 const updateAppointmentListPatient = asyncHandler(async (req, res) => {
   try {
-     if (req.isAuthenticated()) {
+    if (req.isAuthenticated()) {
+      const userGroup = await db.userGroup.findOne({where:{groupName: req.user.user.role}});
+      const authorization = await db.authorizations.findOne({where:{userGroupId: userGroup.id, featId: 5}});
+      if (!authorization.isAccess) { 
+        return res.status(401).json({
+          status: res.statusCode,
+          message: "Unauthorized",
+        });
+      } 
       const { patientId, appointmentListId } = req.body;
       if (!patientId || !appointmentListId) {
         res.status(400).json({
@@ -113,13 +141,13 @@ const updateAppointmentListPatient = asyncHandler(async (req, res) => {
         message: "AppointmentList updated",
         data: appointmentListPatient,
       });
-     } else {
-       res.status(401).json({
-         status: res.statusCode,
-         message: "Unauthorized",
-         data: "",
-       });
-     }
+    } else {
+      res.status(401).json({
+        status: res.statusCode,
+        message: "Unauthorized",
+        data: "",
+      });
+    }
   } catch (err) {
     res.status(500).json({
       status: res.statusCode,
@@ -130,7 +158,15 @@ const updateAppointmentListPatient = asyncHandler(async (req, res) => {
 });
 const deleteAppointmentListPatient = asyncHandler(async (req, res) => {
   try {
-     if (req.isAuthenticated()) {
+    if (req.isAuthenticated()) {
+      const userGroup = await db.userGroup.findOne({where:{groupName: req.user.user.role}});
+      const authorization = await db.authorizations.findOne({where:{userGroupId: userGroup.id, featId: 6}});
+      if (!authorization.isAccess) { 
+        return res.status(401).json({
+          status: res.statusCode,
+          message: "Unauthorized",
+        });
+      } 
       const appointmentListPatient = await db.appointmentListPatient.destroy({
         where: { id: req.params.id },
       });
@@ -139,13 +175,13 @@ const deleteAppointmentListPatient = asyncHandler(async (req, res) => {
         message: "AppointmentList deleted",
         data: appointmentListPatient,
       });
-     } else {
-       res.status(401).json({
-         status: res.statusCode,
-         message: "Unauthorized",
-         data: "",
-       });
-     }
+    } else {
+      res.status(401).json({
+        status: res.statusCode,
+        message: "Unauthorized",
+        data: "",
+      });
+    }
   } catch (err) {
     res.status(500).json({
       status: res.statusCode,
@@ -156,7 +192,15 @@ const deleteAppointmentListPatient = asyncHandler(async (req, res) => {
 });
 const getAppointmentListPatientById = asyncHandler(async (req, res) => {
   try {
-     if (req.isAuthenticated()) {
+    if (req.isAuthenticated()) {
+      const userGroup = await db.userGroup.findOne({where:{groupName: req.user.user.role}});
+      const authorization = await db.authorizations.findOne({where:{userGroupId: userGroup.id, featId: 3}});
+      if (!authorization.isAccess) { 
+        return res.status(401).json({
+          status: res.statusCode,
+          message: "Unauthorized",
+        });
+      } 
       const appointmentList = await db.appointmentListPatient.findOne({
         where: { id: req.params.id },
         include: [
@@ -177,13 +221,13 @@ const getAppointmentListPatientById = asyncHandler(async (req, res) => {
         message: "success",
         data: appointmentList,
       });
-     } else {
-       res.status(401).json({
-         status: res.statusCode,
-         message: "Unauthorized",
-         data: "",
-       });
-     }
+    } else {
+      res.status(401).json({
+        status: res.statusCode,
+        message: "Unauthorized",
+        data: "",
+      });
+    }
   } catch (err) {
     res.status(500).json({
       status: res.statusCode,
